dix: move swaprep.h and swapreq.h into dix
These aren't public includes, so they're better placed in dix.
